Currency_сonverter/requirements.txt
pyperclip==1.8.2
Pillow==9.2.0
